// src/models/DTO/game/websocket/WebsocketErrorEnum.ts
export enum WebsocketErrorEnum {
  RETRY_GENERAL = "RETRY_GENERAL",
  CLOSE_CONNECTION_GENERAL = "CLOSE_CONNECTION_GENERAL",
  CLOSE_CONNECTION_ALREADY_IN_GAME = "CLOSE_CONNECTION_ALREADY_IN_GAME",
  CLOSE_CONNECTION_NO_ACTIVE_GAME = "CLOSE_CONNECTION_NO_ACTIVE_GAME",
  CLOSE_CONNECTION_ALREADY_SUBSCRIBED = "CLOSE_CONNECTION_ALREADY_SUBSCRIBED",
}